refactor resource detail edit/detail selection
Add additional checks for asYamlInit
Refactor checks for showComponent selection
