+++ components/ResourceDetail/index.vue
@@ -1,4 +1,5 @@
 <script>
+import { hasIn } from 'lodash';
 import { cleanForNew } from '@/plugins/steve/normalize';
 import CreateEditView from '@/mixins/create-edit-view';
 import ResourceYaml from '@/components/ResourceYaml';
@@ -67,7 +68,9 @@ export async function defaultAsyncData(ctx, resource) {
 
   const hasCustomDetail = store.getters['type-map/hasCustomDetail'](resource);
   const hasCustomEdit = store.getters['type-map/hasCustomEdit'](resource);
-  const asYamlInit = (route.query[AS_YAML] === _FLAGGED) || (realMode === _VIEW && !hasCustomDetail) || (realMode !== _VIEW && !hasCustomEdit);
+  const asYamlInit = (hasIn(route.query, AS_YAML) && (route.query[AS_YAML] === _FLAGGED)) ||
+   (realMode === _VIEW && !hasCustomDetail && !hasCustomEdit) ||
+    (realMode !== _VIEW && !hasCustomEdit);
   const schema = store.getters['cluster/schemaFor'](resource);
   const schemas = store.getters['cluster/all'](SCHEMA);
 
@@ -187,11 +190,14 @@ export default {
 
   data() {
     // asYamlInit is taken from route query and passed as prop from _id page; asYaml is saved in local data to be manipulated by Masthead
-    const asYaml = this.asYamlInit;
+    const {
+      asYamlInit: asYaml,
+      value: currentValue,
+    } = this;
 
     return {
       asYaml,
-      currentValue:            this.value,
+      currentValue,
       detailComponent:         this.$store.getters['type-map/importDetail'](this.resource),
       editComponent:           this.$store.getters['type-map/importEdit'](this.resource),
     };
@@ -230,7 +236,7 @@ export default {
     showComponent() {
       if ( this.isView && this.hasCustomDetail ) {
         return this.detailComponent;
-      } else if ( !this.isView && this.hasCustomEdit ) {
+      } else if ( !this.asYaml && this.hasCustomEdit ) {
         return this.editComponent;
       }
 
@@ -267,7 +273,21 @@ export default {
       :as-yaml.sync="asYaml"
       :has-detail="hasCustomDetail"
     />
-    <template v-if="asYaml">
+    <template v-if="(hasCustomEdit || hasCustomDetail) && showComponent">
+      <component
+        :is="showComponent"
+        v-model="model"
+        :original-value="originalModel"
+        :done-route="doneRoute"
+        :done-params="doneParams"
+        :mode="mode"
+        :real-mode="realMode"
+        :value="model"
+        :obj="model"
+        :yaml="yaml"
+      />
+    </template>
+    <template v-else>
       <div v-if="!isView">
         <button class="btn btn-sm role-primary" @click="readFromFile">
           Read from file
@@ -280,20 +300,6 @@ export default {
         :yaml="yaml"
         :offer-preview="offerPreview"
         :done-route="doneRoute"
-      />
-    </template>
-    <template v-else>
-      <component
-        :is="showComponent"
-        v-model="model"
-        :original-value="originalModel"
-        :done-route="doneRoute"
-        :done-params="doneParams"
-        :mode="mode"
-        :real-mode="realMode"
-        :value="model"
-        :obj="model"
-        :yaml="yaml"
       />
     </template>
   </div>
